Added User Repository and restructured file system for back end
