+++ src/main/java/com/scottlogic/librarygradproject/Entities/Book.java
@@ -1,4 +1,4 @@
-package com.scottlogic.librarygradproject;
+package com.scottlogic.librarygradproject.Entities;
 
 import lombok.Builder;
 import javax.persistence.Entity;
