+++ src/main/java/com/scottlogic/librarygradproject/Exceptions/BookNotFoundException.java
@@ -1,4 +1,4 @@
-package com.scottlogic.librarygradproject;
+package com.scottlogic.librarygradproject.Exceptions;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
