+++ src/main/java/com/scottlogic/librarygradproject/Repositories/BookRepository.java
@@ -1,5 +1,6 @@
-package com.scottlogic.librarygradproject;
+package com.scottlogic.librarygradproject.Repositories;
 
+import com.scottlogic.librarygradproject.Entities.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
