+++ src/test/java/com/scottlogic/librarygradproject/BookServiceTest.java
@@ -1,5 +1,8 @@
 package com.scottlogic.librarygradproject;
 
+import com.scottlogic.librarygradproject.Entities.Book;
+import com.scottlogic.librarygradproject.Exceptions.IncorrectBookFormatException;
+import com.scottlogic.librarygradproject.Services.BookService;
 import org.apache.commons.lang3.StringUtils;
 import org.junit.Test;
 import org.junit.runner.RunWith;
